packages/monai-server/requirements.txt
# ── Core (required) ──────────────────────────────────────────────────
monai>=1.3.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
Pillow>=10.0.0
pydicom>=2.4.0
highdicom>=0.22.0
pylibjpeg>=2.0.0
pylibjpeg-libjpeg>=2.1.0
pylibjpeg-openjpeg>=2.1.0
einops>=0.7.0
httpx>=0.27.0
requests>=2.31.0
pydantic>=2.0.0
# ── Pre-trained weights (free, MIT) ─────────────────────────────────
torchxrayvision>=1.2.0
torchvision>=0.15.0
